gen-pull: Automatically determine start and end tags
In case we submit multiple pull requests per merge window, no longer
hardcode the start and end tags that we are going to submit, but
determine them by looking at the intermediate tags as well.

This involves a bit of code refactor and makes find_baseline_tag()
return start and end.

Signed-off-by: Florian Fainelli <f.fainelli@gmail.com>